refactor!: Moved `ObjectIO` to the `utils` package.
