+++ src/main/java/com/gersonfaneto/yams/dao/entities/user/UserDiskDAO.java
@@ -1,10 +1,11 @@
 package com.gersonfaneto.yams.dao.entities.user;
 
 import com.gersonfaneto.yams.dao.CRUD;
-import com.gersonfaneto.yams.dao.ObjectIO;
 import com.gersonfaneto.yams.dao.Persist;
 import com.gersonfaneto.yams.models.entities.user.User;
 import com.gersonfaneto.yams.utils.Generators;
+import com.gersonfaneto.yams.utils.ObjectIO;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
