+++ src/main/java/com/gersonfaneto/yams/dao/services/service/ServiceDiskDAO.java
@@ -1,11 +1,12 @@
 package com.gersonfaneto.yams.dao.services.service;
 
 import com.gersonfaneto.yams.dao.CRUD;
-import com.gersonfaneto.yams.dao.ObjectIO;
 import com.gersonfaneto.yams.dao.Persist;
 import com.gersonfaneto.yams.models.services.service.Service;
 import com.gersonfaneto.yams.models.services.service.ServiceType;
 import com.gersonfaneto.yams.utils.Generators;
+import com.gersonfaneto.yams.utils.ObjectIO;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
